Split elastic.py into a client and the interface the rest of our app uses.
Also removes ELASTICSEARCH_URL from the database configuration because it doesn't make sense to change the elastic node(s) you're point to while you're already running. In almost every case, it's going to be set once and rarely changed.
